+++ assets/js/visualmarkdownfield.js
@@ -1,7 +1,7 @@
 /**
  * Visual Markdown Editor Field for Kirby 2
  *
- * @version   1.3.4
+ * @version   1.5.0
  * @author    Jonas Döbertin <hello@jd-powered.net>
  * @copyright Jonas Döbertin <hello@jd-powered.net>
  * @link      https://github.com/JonasDoebertin/kirby-visual-markdown
@@ -13,27 +13,29 @@
  *
  * @since 1.0.0
  */
-var VisualMarkdownField = function($, $field) {
+var VisualMarkdownField = function ($, $field) {
     'use strict';
 
     var self = this;
 
-    this.$field     = $field;
+    this.$field = $field;
     this.$draggable = $('.sidebar').find('.draggable');
-    this.$wrapper   = $field.closest('.markdownfield-wrapper');
-    this.$toolbar   = null;
-    this.$editor    = null;
+    this.$wrapper = $field.closest('.markdownfield-wrapper');
+    this.$toolbar = null;
+    this.$editor = null;
 
     this.editor = null;
 
+    this.inAction = false;
     this.isFixed = false;
+    this.isFocused = false;
     this.isFullscreen = false;
 
     this.options = {
-        toolbar:   $field.data('toolbar'),
-        header1:   $field.data('header1'),
-        header2:   $field.data('header2'),
-        tools:     $field.data('tools').split(','),
+        toolbar: $field.data('toolbar'),
+        header1: $field.data('header1'),
+        header2: $field.data('header2'),
+        tools: $field.data('tools').split(','),
         kirbytext: $field.data('kirbytext')
     };
 
@@ -42,20 +44,20 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.0.0
      */
-    this.init = function() {
+    this.init = function () {
 
         /*
             Initialize VisualMarkdownEditor
          */
-        self.editor = new VisualMarkdownEditor($, self.$field, self.options);
+        self.editor = new VisualMarkdownEditor($, self, self.$field, self.options);
 
         /*
             Store some references to the underlying codemirror instance
             our field instance and the toolbar and editor elements
          */
         self.codemirror = self.editor.getCodeMirrorInstance();
-        self.$toolbar   = self.$field.siblings('.visualmarkdown-toolbar');
-        self.$editor    = self.$field.siblings('.CodeMirror');
+        self.$toolbar = self.$field.siblings('.visualmarkdown-toolbar');
+        self.$editor = self.$field.siblings('.CodeMirror');
 
         /*
             Set up change event handler
@@ -81,8 +83,8 @@ var VisualMarkdownField = function($, $field) {
          */
         self.$wrapper.droppable({
             hoverClass: 'markdownfield-wrapper-over',
-            accept:     self.draggable,
-            drop:       function(event, element) {
+            accept: self.draggable,
+            drop: function (event, element) {
                 self.editor.insert(element.draggable.data('text'));
             }
         });
@@ -92,15 +94,16 @@ var VisualMarkdownField = function($, $field) {
          *
          * @since 1.4.0
          */
-        $(window).scroll(function() {
+        $('.mainbar').scroll(function () {
 
             /**
              * Switch to fixed toolbar, if
              * - the fullscreen mode isn't enabled
              * - the toolbar isn't fixed already
              * - the scroll position is within the fields wrapper
+             * - the field is focused
              */
-            if (!self.isFullscreen && !self.isFixed && self.scrollTopWithinWrapper()) {
+            if (!self.isFullscreen && !self.isFixed && self.scrollTopWithinWrapper() && self.isFocused) {
                 self.enableFixedToolbar();
             }
 
@@ -122,9 +125,16 @@ var VisualMarkdownField = function($, $field) {
          *
          * @since 1.0.0
          */
-        self.$field.bind('destroyed', function() {
+        self.$field.bind('destroyed', function () {
             self.deactivate();
         });
+
+        /**
+         * FIX: Add an indicator that this field is already initialized to
+         * prevent double initialization of a Visual MArkdown Editor field.
+         * See: https://github.com/JonasDoebertin/kirby-visual-markdown/issues/61
+         */
+        self.$field.data('initialized', true);
 
     };
 
@@ -133,7 +143,7 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.0.0
      */
-    this.updateStorage = function() {
+    this.updateStorage = function () {
         self.$field.text(self.codemirror.getValue());
     };
 
@@ -142,7 +152,7 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.0.0
      */
-    this.deactivate = function() {
+    this.deactivate = function () {
         self.updateStorage();
         self.editor.deactivate();
     };
@@ -152,15 +162,15 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.0.1
      */
-    this.changeFullscreenModeHandler = function() {
+    this.changeFullscreenModeHandler = function () {
 
         // Add indication class if fullscreen mode was entered
-        if(screenfull.isFullscreen && (screenfull.element === self.$wrapper.get(0))) {
+        if (screenfull.isFullscreen && (screenfull.element === self.$wrapper.get(0))) {
             self.attachFullscreenStyles();
         }
 
         // Remove indicator class if fullscreen mode was exited
-        if(!screenfull.isFullscreen) {
+        if (!screenfull.isFullscreen) {
             self.detachFullscreenStyles();
         }
 
@@ -171,8 +181,12 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.0.0
      */
-    this.attachFocusStyles = function() {
+    this.attachFocusStyles = function () {
+        self.isFocused = true;
         self.$wrapper.addClass('markdownfield-wrapper-focused');
+        if (!self.isFullscreen && self.scrollTopWithinWrapper()) {
+            self.enableFixedToolbar();
+        }
     };
 
     /**
@@ -180,8 +194,10 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.0.0
      */
-    this.detachFocusStyles = function() {
+    this.detachFocusStyles = function () {
+        self.isFocused = false;
         self.$wrapper.removeClass('markdownfield-wrapper-focused');
+        self.disableFixedToolbar();
     };
 
     /**
@@ -189,7 +205,7 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.0.1
      */
-    this.attachFullscreenStyles = function() {
+    this.attachFullscreenStyles = function () {
         self.isFullscreen = true;
         self.disableFixedToolbar();
         self.$wrapper.addClass('markdownfield-wrapper-fullscreen');
@@ -200,7 +216,7 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.0.1
      */
-    this.detachFullscreenStyles = function() {
+    this.detachFullscreenStyles = function () {
         self.isFullscreen = false;
         self.$wrapper.removeClass('markdownfield-wrapper-fullscreen');
     };
@@ -211,10 +227,11 @@ var VisualMarkdownField = function($, $field) {
      * @since 1.4.0
      * @return boolean
      */
-    this.scrollTopWithinWrapper = function() {
-        var $document = $(document);
-        return (($document.scrollTop() >= self.$wrapper.offset().top)
-                && ($document.scrollTop() < self.$wrapper.offset().top + self.$wrapper.outerHeight() - self.$toolbar.outerHeight()));
+    this.scrollTopWithinWrapper = function () {
+        var topOffset    = self.$wrapper.offset().top + 2,
+            bottomOffset = self.$wrapper.offset().top + self.$wrapper.outerHeight() - self.$toolbar.outerHeight() - 2;
+
+        return ((topOffset <= 48) && (bottomOffset >= 48));
     };
 
     /**
@@ -222,10 +239,10 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.4.0
      */
-    this.enableFixedToolbar = function() {
+    this.enableFixedToolbar = function () {
         self.isFixed = true;
         self.$toolbar.addClass('visualmarkdown-toolbar-fixed')
-                     .css('max-width', self.$wrapper.width());
+            .css('max-width', self.$wrapper.width());
         self.$wrapper.css('padding-top', self.$toolbar.outerHeight());
     };
 
@@ -234,11 +251,13 @@ var VisualMarkdownField = function($, $field) {
      *
      * @since 1.4.0
      */
-    this.disableFixedToolbar = function() {
-        self.isFixed = false;
-        self.$toolbar.removeClass('visualmarkdown-toolbar-fixed')
-                     .css('max-width', '');
-        self.$wrapper.css('padding-top', 0);
+    this.disableFixedToolbar = function () {
+        if (!self.inAction) {
+            self.isFixed = false;
+            self.$toolbar.removeClass('visualmarkdown-toolbar-fixed')
+                .css('max-width', '');
+            self.$wrapper.css('padding-top', 0);
+        }
     };
 
     /**
@@ -248,7 +267,7 @@ var VisualMarkdownField = function($, $field) {
 
 };
 
-(function($) {
+(function ($) {
     'use strict';
 
     /**
@@ -257,8 +276,8 @@ var VisualMarkdownField = function($, $field) {
      * @since 1.0.0
      */
     $.event.special.destroyed = {
-        remove: function(event) {
-            if(event.handler) {
+        remove: function (event) {
+            if (event.handler) {
                 event.handler.apply(this, arguments);
             }
         }
@@ -273,8 +292,10 @@ var VisualMarkdownField = function($, $field) {
      * @see https://github.com/getkirby/panel/issues/228#issuecomment-58379016
      * @since 1.0.0
      */
-    $.fn.markdownfield = function() {
-        return new VisualMarkdownField($, this);
+    $.fn.markdownfield = function () {
+        if (!this.data('initialized')) {
+            return new VisualMarkdownField($, this);
+        }
     };
 
 })(jQuery);