+++ languages/tr.php
@@ -1,0 +1,55 @@
+<?php
+
+/**
+ * @language Turkish
+ * @direction ltr
+ * @author Alkın Kasap <admin@alkinkasap.net>
+ */
+
+return [
+
+    /* Main toolbar icons */
+    'action.tooltip.header1'       => 'Birinci Başlık',
+    'action.tooltip.header2'       => 'İkinci Başlık',
+    'action.name.header2'          => 'h',
+    'action.tooltip.bold'          => 'Kalın',
+    'action.tooltip.italic'        => 'Eğik',
+    'action.tooltip.strikethrough' => 'Orta Çizgili',
+    'action.tooltip.blockquote'    => 'Alıntı',
+    'action.tooltip.unorderedList' => 'Sırasız Liste',
+    'action.tooltip.orderedList'   => 'Sıralı Liste',
+    'action.tooltip.link'          => 'Bağlantı Ekle',
+    'action.tooltip.email'         => 'Eposta Bağlantısı Ekle',
+    'action.tooltip.image'         => 'Resim Ekle',
+    'action.tooltip.line'          => 'Orta Çizgi Ekle',
+    'action.tooltip.fullscreen'    => 'Tam Ekran Modu',
+
+    /* Help icon subitems */
+    'action.tooltip.shortcutsModal' => 'Bütün Klavye Kısayolları',
+    'action.name.shortcutsModal'    => 'Klavye Kısayolları',
+    'action.tooltip.markdownLink'   => 'Markdown Hakkında Bilmeniz Gerekenler',
+    'action.name.markdownLink'      => 'Markdown Yardım',
+    'action.tooltip.kirbytextLink'  => 'Kirbytext ile Yazım',
+    'action.name.kirbytextLink'     => 'Kirbytext Yardım',
+    'action.tooltip.issuesLink'     => 'Githubta Soru Aç',
+    'action.name.issuesLink'        => 'Hata Bildir',
+    'action.tooltip.licenseLink'    => 'Moral Lisans Alarak Yapımcıyı Destekleyin',
+    'action.name.licenseLink'       => 'Lisans Al',
+
+    /* Shortcuts modal texts */
+    'modal.shortcuts.title'                => 'Klavye Kısayolları',
+    'modal.shortcuts.heading.action'       => 'Action',
+    'modal.shortcuts.heading.shortcut'     => 'Shortcut',
+    'modal.shortcuts.action.heading'       => 'Başlık',
+    'modal.shortcuts.action.subheading'    => 'Altbaşlık',
+    'modal.shortcuts.action.bold'          => 'Kalın',
+    'modal.shortcuts.action.italic'        => 'Eğik',
+    'modal.shortcuts.action.strikethrough' => 'Orta Çizgili',
+    'modal.shortcuts.action.blockquote'    => 'Alıntı',
+    'modal.shortcuts.action.unorderedList' => 'Sırasız Liste',
+    'modal.shortcuts.action.orderedList'   => 'Sıralı Liste',
+    'modal.shortcuts.action.image'         => 'Resim',
+    'modal.shortcuts.action.link'          => 'Bağlantı',
+    'modal.shortcuts.action.email'         => 'Eposta Bağlantısı',
+
+];